Created Interaction Events and Interactable Components
These classes are the core of a new system to handle interactions between player pawns and different objects in the game world. InteractionEvents originate from an interactor (e.g. player pawn) and can do their own functionality. If the interactee is an InteractableComponent, it can implement its own functionality that occurs after the InteractionEvent.